+++ compose.yaml
@@ -1,15 +1,36 @@
 services:
+  worker:
+    build:
+      context: .
+      dockerfile: Dockerfile.web
+    ports:
+      - "8100:8100"
+    environment:
+      - DATABASE_URL=postgres://igorreadwrite:readwrite@postgres:5432/igor?sslmode=disable&search_path=igor
+      - REDIS_URL=redis:6379
+      - ADDR=0.0.0.0:8100
+      - DEBUG=true
+    depends_on:
+      - postgres
+      - redis
+    deploy:
+      mode: replicated
+      replicas: 1
   postgres:
     image: "postgres:16"
     ports:
       - "5432:5432"
     environment:
       - POSTGRES_PASSWORD=test
-#  opensearch:
-#    image: "opensearchproject/opensearch:latest"
-#    ports:
-#      - "9200:9200"
-#  dashboards:
-#    image: "opensearchproject/opensearch-dashboards:latest"
-#    ports:
-#      - "5601:8082"
+  redis:
+    image: "redis:alpine"
+    ports:
+      - "6379:6379"
+  asynqmon:
+    image: "hibiken/asynqmon"
+    ports:
+      - "8080:8080"
+    environment:
+      - REDIS_ADDR=redis:6379
+    depends_on:
+      - redis
