--- experimental/README.md
@@ -1,1 +1,0 @@
-Nothing under /experimental is necessarily worked or recommended :P
