--- pkg/api/code/api.go
@@ -1,7 +1,0 @@
-// code API module is intended for those who wish to use Igor by directly compiling against it.
-//
-// Ie. it provides an interface similar to systems like Asynq or Tasque.
-//
-// If you need to call Igor from other languages then you'll want another API module (eg. HTTP, gRPC, etc)
-// or add your own.
-package code
